+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## New Features
 
+ - PR #106 Added multi-GPU initialization
+
 ## Improvements
 
 - PR #161 Use `std::atexit` to finalize RMM after Python interpreter shutdown
@@ -13,7 +15,7 @@
 - PR #170 Always build librmm and rmm, but conditionally upload based on CUDA / Python version
 
 
-# RMM 0.10.0 (Date TBD)
+# RMM 0.10.0 (16 Oct 2019)
 
 ## New Features
 
@@ -40,7 +42,7 @@
  - PR #157 Check if initialized before freeing for Numba finalizer and use `weakref` instead of `atexit`
 
 
-# RMM 0.9.0 (21 August 2019)
+# RMM 0.9.0 (21 Aug 2019)
 
 ## New Features